rowid are int64_t

+++ libpkg/pkg_private.h
@@ -25,7 +25,7 @@ struct pkg {
 	struct array exec;
 	struct array options;
 	int flags;
-	int rowid;
+	int64_t rowid;
 	pkg_t type;
 };
